requirements_fast.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.5.1
PyMuPDF==1.23.8
chromadb==0.4.15
sentence-transformers==2.2.2
torch==2.0.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.30.2
llama-index-core==0.10.1
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-chroma==0.1.7
